rename template files to consistent with configuration file names
