+++ kong/cmd/utils/prefix_handler.lua
@@ -2,9 +2,9 @@ local default_nginx_template = require "kong.templates.nginx"
 local kong_nginx_template = require "kong.templates.nginx_kong"
 local kong_nginx_gui_include_template = require "kong.templates.nginx_kong_gui_include"
 local kong_nginx_stream_template = require "kong.templates.nginx_kong_stream"
-local nginx_main_inject_template = require "kong.templates.nginx_main_inject"
-local nginx_http_inject_template = require "kong.templates.nginx_http_inject"
-local nginx_stream_inject_template = require "kong.templates.nginx_stream_inject"
+local nginx_main_inject_template = require "kong.templates.nginx_inject"
+local nginx_http_inject_template = require "kong.templates.nginx_kong_inject"
+local nginx_stream_inject_template = require "kong.templates.nginx_kong_stream_inject"
 local system_constants = require "lua_system_constants"
 local process_secrets = require "kong.cmd.utils.process_secrets"
 local openssl_bignum = require "resty.openssl.bn"
