+++ kong/templates/nginx_inject.lua
