+++ kong/templates/nginx_kong_inject.lua
