+++ kong/templates/nginx_kong_stream_inject.lua
